Add sbt-tpolecat and fix errors
