+++ project/plugins.sbt
@@ -1,0 +1,1 @@
+addSbtPlugin("io.github.davidgregory084" % "sbt-tpolecat" % "0.1.20")
