--- src/main/scala/zone/slice/chaos/publish.scala
@@ -1,62 +1,0 @@
-package zone.slice.chaos
-
-import fs2.Stream
-import fs2.concurrent.Topic
-import cats.{Applicative, Functor}
-import cats.syntax.all._
-import cats.effect.std.Console
-import org.http4s.{Uri, Method}
-import org.http4s.client.Client
-import org.http4s.{Request, EntityEncoder, EntityDecoder, EntityBody}
-
-package object publish {
-
-  /** Publishing operations that publishers can perform. */
-  trait Publish[F[_]] {
-    def output(text: String): F[Unit]
-
-    def request[A](request: Request[F])(implicit ev: EntityDecoder[F, A]): F[A]
-
-    def get[A](from: Uri)(implicit ev: EntityDecoder[F, A]): F[A] =
-      this.request(Request[F](uri = from))
-
-    def post[I, O](to: Uri, body: I)(implicit
-        decoder: EntityDecoder[F, O],
-        encoder: EntityEncoder[F, I],
-    ): F[O] =
-      this.request(Request[F](method = Method.POST, uri = to).withEntity(body))
-  }
-
-  object Publish {
-    def make[F[_]](console: Console[F], client: Client[F]): Publish[F] =
-      new Publish[F] {
-        def output(text: String): F[Unit] =
-          console.errorln(text)
-        def request[A](request: Request[F])(implicit
-            decoder: EntityDecoder[F, A],
-        ): F[A] =
-          client.expect(request)
-      }
-  }
-
-  /** A function that publishes something. */
-  type Publisher[F[_], -A] = (A, Publish[F]) => F[Unit]
-
-  implicit class PublisherOps[F[_], A](publisher: Publisher[F, A]) {
-
-    /** Conditionalizes a publisher. */
-    def when(cond: A => Boolean)(implicit
-        ev: Applicative[F],
-    ): Publisher[F, A] =
-      (a, p) => publisher(a, p).whenA(cond(a))
-  }
-
-  /** Forward things from a topic into a publisher. */
-  def subscribe[F[_]: Functor, A](topic: Topic[F, A], f: Publisher[F, A])(
-      implicit publish: Publish[F],
-  ): Stream[F, Nothing] =
-    topic
-      .subscribe(0)
-      .evalTap(f(_, publish))
-      .drain
-}
